requirements.txt
Flask==2.3.2
injector==0.20.1
python-dotenv==1.0.0
openai==0.27.6
tiktoken==0.3.3
tqdm==4.65.0
streamlit